// Proyectos/Aldeas.Web/ClientApp/src/app/models/comunes.ts
// Generated by https://quicktype.io

export interface CentrosResponse {
    centrosCostos: CentrosCosto[];
    subCentro:     SubCentro[];
}

export interface CentrosCosto {
    codigoCentro: number;
    nombre:       string;
}

export interface SubCentro {
    codigoCentro:    number;
    codigoSubCentro: number;
    nombre:          string;
}


export interface Respuesta{
    id:    number;
    code: number;
    status:  string;
    message: string;
}


export interface SelectGlobal {
    value: string;
    viewValue: string;
  }
  